:sparkles: GUI with context menu on top with "Open ROM", "Quit" and "Zoom In/Out"
