+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
 
 [dependencies]
 minifb = "0.28"
-eframe = "0.28"
-egui = "0.28"
-rfd = "0.14" # for file dialogs
+eframe = "0.33.2"
+egui = "0.33.2"
+egui_extras = "0.33.2"
+rfd = "0.14"
